+++ perllib/FixMyStreet/TestMech.pm
@@ -706,7 +706,7 @@ sub create_problems_for_body {
             latitude           => '51.5016605453401',
             longitude          => '-0.142497580865087',
             user_id            => $user->id,
-            photo              => $mech->get_photo_data,
+            photo              => '74e3362283b6ef0c48686fb0e161da4043bbcc97.jpeg',
         };
 
         my %report_params = ( %$default_params, %$params );
@@ -719,15 +719,6 @@ sub create_problems_for_body {
     }
 
     return @problems;
-}
-
-sub get_photo_data {
-    my $mech = shift;
-    return $mech->{sample_photo} ||= do {
-        my $sample_file = FixMyStreet->path_to( 't/app/controller/sample.jpg' );
-        $mech->builder->ok( -f "$sample_file", "sample file $sample_file exists" );
-        $sample_file->slurp(iomode => '<:raw');
-    };
 }
 
 sub create_comment_for_problem {