+++ CHANGELOG.md
@@ -7,9 +7,10 @@
         - Allow cobrands to limit contact form to abuse reports only
     - Front end improvements:
         - Clearer relocation options while you’re reporting a problem #2238
-    - Admin improvements
+    - Admin improvements:
         - List number of alerts on report page #669
         - viewing and managing of user alerts in admin #676
+        - Allow moderation to potentially change photos. #2291
     - Bugfixes:
         - Add perl 5.26/5.28 support.
         - Fix subcategory issues when visiting /report/new directly #2276